# contracts/api-contracts/src/main/resources/openapi/signalharvester-v1.yaml
openapi: 3.1.0
info:
  title: SignalHarvester Backend API
  version: 0.1.0
  description: Public HTTP contract for SignalHarvester configuration, results, event observation, processing-flow, and operational
    APIs.
...
security:
- cookieAuth: []
- bearerAuth: []
paths:
  /api/v1/sources:
    get:
      tags:
      - Sources
      operationId: listSources
      summary: List configured external sources
      responses:
        '200':
          description: Configured sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Source'
    post:
      tags:
      - Sources
      operationId: createSource
      summary: Create an external source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceUpsertRequest'
      responses:
        '201':
          description: Source created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '400':
          description: Invalid source configuration
  /api/v1/sources/{sourceId}:
    parameters:
    - $ref: '#/components/parameters/SourceId'
    get:
      tags:
      - Sources
      operationId: getSource
      summary: Get a configured external source
      responses:
        '200':
          description: Configured source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          description: Source not found
    put:
      tags:
      - Sources
      operationId: updateSource
      summary: Replace an external source configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceUpsertRequest'
      responses:
        '200':
          description: Source updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '400':
          description: Invalid source configuration
        '404':
          description: Source not found
    delete:
      tags:
      - Sources
      operationId: deleteSource
      summary: Delete an external source
      responses:
        '204':
          description: Source deleted
        '404':
          description: Source not found
        '409':
          description: Source is referenced by a monitoring profile
...
  /api/v1/admin/analysis/dead-letters/{partition}/{offset}:
    parameters:
    - $ref: '#/components/parameters/DeadLetterPartition'
    - $ref: '#/components/parameters/DeadLetterOffset'
    get:
      tags: [Dead Letter Recovery]
      operationId: inspectAnalysisDeadLetter
      summary: Inspect one Analysis dead-letter record
      description: Reads one real Analysis DLQ record by Kafka position without exposing its serialized source payload.
      responses:
        '200':
          description: Sanitized Analysis dead-letter metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadLetterInspection'
        '400': { description: Invalid DLQ position }
        '401': { description: Authentication required }
        '403': { description: ADMIN role required }
        '404': { description: DLQ record not found }
        '422': { description: DLQ record is invalid for Analysis recovery }
        '429': { description: Analysis recovery concurrency limit reached }
        '503': { description: Kafka unavailable }
  /api/v1/admin/analysis/dead-letters/{partition}/{offset}/replay:
    parameters:
    - $ref: '#/components/parameters/DeadLetterPartition'
    - $ref: '#/components/parameters/DeadLetterOffset'
    post:
      tags: [Dead Letter Recovery]
      operationId: replayAnalysisDeadLetter
      summary: Replay one confirmed Analysis dead-letter record
      description: Reprocesses the stored original key and payload through the Analysis-owned decoder/application path without republishing the shared source event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeadLetterReplayRequest'
      responses:
        '200':
          description: Sanitized metadata for the replayed Analysis dead-letter record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadLetterInspection'
        '400': { description: Invalid request or DLQ position }
        '401': { description: Authentication required }
        '403': { description: ADMIN role required or CSRF validation failed }
        '404': { description: DLQ record not found }
        '409': { description: Confirmation mismatch or Analysis replay failed }
        '422': { description: DLQ record or stored source record is invalid for Analysis recovery }
        '429': { description: Analysis recovery concurrency limit reached }
        '503': { description: Kafka unavailable }
  /api/v1/admin/results/dead-letters/{partition}/{offset}:
    parameters:
    - $ref: '#/components/parameters/DeadLetterPartition'
    - $ref: '#/components/parameters/DeadLetterOffset'
    get:
      tags: [Dead Letter Recovery]
      operationId: inspectResultsDeadLetter
      summary: Inspect one Results dead-letter record
      description: Reads one real Results DLQ record by Kafka position without exposing its serialized source payload.
      responses:
        '200':
          description: Sanitized Results dead-letter metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadLetterInspection'
        '400': { description: Invalid DLQ position }
        '401': { description: Authentication required }
        '403': { description: ADMIN role required }
        '404': { description: DLQ record not found }
        '422': { description: DLQ record is invalid for Results recovery }
        '429': { description: Results recovery concurrency limit reached }
        '503': { description: Kafka unavailable }
  /api/v1/admin/results/dead-letters/{partition}/{offset}/replay:
    parameters:
    - $ref: '#/components/parameters/DeadLetterPartition'
    - $ref: '#/components/parameters/DeadLetterOffset'
    post:
      tags: [Dead Letter Recovery]
      operationId: replayResultsDeadLetter
      summary: Replay one confirmed Results dead-letter record
      description: Reprojects the stored original key and payload through the Results-owned decoder/application path without republishing the shared Analysis event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeadLetterReplayRequest'
      responses:
        '200':
          description: Sanitized metadata for the replayed Results dead-letter record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadLetterInspection'
        '400': { description: Invalid request or DLQ position }
        '401': { description: Authentication required }
        '403': { description: ADMIN role required or CSRF validation failed }
        '404': { description: DLQ record not found }
        '409': { description: Confirmation mismatch or Results replay failed }
        '422': { description: DLQ record or stored source record is invalid for Results recovery }
        '429': { description: Results recovery concurrency limit reached }
        '503': { description: Kafka unavailable }
  /api/v1/admin/event-observation/dead-letters/{partition}/{offset}:
    parameters:
    - $ref: '#/components/parameters/DeadLetterPartition'
    - $ref: '#/components/parameters/DeadLetterOffset'
    get:
      tags: [Dead Letter Recovery]
      operationId: inspectEventObservationDeadLetter
      summary: Inspect one Event Observation dead-letter record
      description: Reads one real Event Observation DLQ record by Kafka position without exposing its serialized source payload.
      responses:
        '200':
          description: Sanitized Event Observation dead-letter metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadLetterInspection'
        '400': { description: Invalid DLQ position }
        '401': { description: Authentication required }
        '403': { description: ADMIN role required }
        '404': { description: DLQ record not found }
        '422': { description: DLQ record is invalid for Event Observation recovery }
        '429': { description: Event Observation recovery concurrency limit reached }
        '503': { description: Kafka unavailable }
  /api/v1/admin/event-observation/dead-letters/{partition}/{offset}/replay:
    parameters:
    - $ref: '#/components/parameters/DeadLetterPartition'
    - $ref: '#/components/parameters/DeadLetterOffset'
    post:
      tags: [Dead Letter Recovery]
      operationId: replayEventObservationDeadLetter
      summary: Replay one confirmed Event Observation dead-letter record
      description: Re-records the stored original key and payload through Event Observation only, preserving original source transport metadata and without republishing the shared source event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeadLetterReplayRequest'
      responses:
        '200':
          description: Sanitized metadata for the replayed Event Observation dead-letter record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadLetterInspection'
        '400': { description: Invalid request or DLQ position }
        '401': { description: Authentication required }
        '403': { description: ADMIN role required or CSRF validation failed }
        '404': { description: DLQ record not found }
        '409': { description: Confirmation mismatch or Event Observation replay failed }
        '422': { description: DLQ record or stored source record is invalid for Event Observation recovery }
        '429': { description: Event Observation recovery concurrency limit reached }
        '503': { description: Kafka unavailable }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SIGNALHARVESTER_AUTH
      description: HttpOnly JWT cookie used by browser REST and native EventSource requests.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer transport for approved non-browser clients.
  parameters:
    DeadLetterPartition:
      name: partition
      in: path
      required: true
      description: Kafka partition of the owner-specific dead-letter topic.
      schema:
        type: integer
        minimum: 0
    DeadLetterOffset:
      name: offset
      in: path
      required: true
      description: Kafka offset of the owner-specific dead-letter topic.
      schema:
        type: integer
        format: int64
        minimum: 0
    SourceId:
      name: sourceId
      in: path
      required: true
      description: Stable source identifier.
      schema:
        type: string
        format: uuid
  schemas:
    DeadLetterReplayRequest:
      type: object
      required: [expectedDeadLetterId]
      properties:
        expectedDeadLetterId:
          type: string
          minLength: 1
          maxLength: 1024
          pattern: .*\S.*
          description: Exact dead-letter identifier returned by inspection and required as an explicit replay confirmation.
    DeadLetterInspection:
      type: object
      required:
      - deadLetterId
      - consumer
      - consumerGroup
      - deadLetterTopic
      - deadLetterPartition
      - deadLetterOffset
      - sourceTopic
      - sourcePartition
      - sourceOffset
      - sourceKey
      - failureType
      - failureMessage
      - attempts
      - retryable
      - sourcePayloadBytes
      properties:
        deadLetterId:
          type: string
        consumer:
          type: string
        consumerGroup:
          type: string
        deadLetterTopic:
          type: string
        deadLetterPartition:
          type: integer
          minimum: 0
        deadLetterOffset:
          type: integer
          format: int64
          minimum: 0
        sourceTopic:
          type: string
        sourcePartition:
          type: integer
          minimum: 0
        sourceOffset:
          type: integer
          format: int64
          minimum: 0
        sourceKey:
          type: string
          description: Original Kafka key, or an empty string when the source record had no usable key.
        failureType:
          type: string
        failureMessage:
          type: string
        attempts:
          type: integer
          minimum: 1
        retryable:
          type: boolean
        sourcePayloadBytes:
          type: integer
          minimum: 0
          description: Serialized source payload size; the payload bytes themselves are never returned by the administrative API.
...
    SourceType:
      type: string
      enum:
      - REST
      - RSS
      - HTML
    Source:
      type: object
      required:
      - id
      - name
      - type
      - location
      - enabled
      - settings
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          pattern: .*\S.*
        type:
          $ref: '#/components/schemas/SourceType'
        location:
          type: string
          format: uri
          pattern: ^[Hh][Tt][Tt][Pp][Ss]?://(?![^/?#]*@)(?:\[[0-9A-Fa-f:.]+\]|[^/?#:\s]+)(?::\d+)?(?:[/?][^#\s]*)?$
          description: Absolute HTTP(S) source URL without embedded credentials or
            a fragment.
        enabled:
          type: boolean
        settings:
          type: object
          description: Source-type-specific string configuration interpreted by the owning collector. REST JSON extraction uses json.* JSON Pointer settings; HTML extraction uses html.* CSS selector settings.
          additionalProperties:
            type: string
    SourceUpsertRequest:
      type: object
      required:
      - name
      - type
      - location
      properties:
        name:
          type: string
          minLength: 1
          pattern: .*\S.*
        type:
          $ref: '#/components/schemas/SourceType'
        location:
          type: string
          format: uri
          pattern: ^[Hh][Tt][Tt][Pp][Ss]?://(?![^/?#]*@)(?:\[[0-9A-Fa-f:.]+\]|[^/?#:\s]+)(?::\d+)?(?:[/?][^#\s]*)?$
          description: Absolute HTTP(S) source URL without embedded credentials or
            a fragment.
        enabled:
          type: boolean
          default: false
          description: Whether collection is enabled. Omitted requests default to false.
        settings:
          type: object
          default: {}
          description: Source-type-specific string configuration interpreted by the owning collector. REST JSON extraction uses json.* JSON Pointer settings; HTML extraction uses html.* CSS selector settings.
          additionalProperties:
            type: string
...